splitting off navigator
