+++ core/include/tools/planar_intersector.hpp
@@ -93,11 +93,11 @@ namespace detray
             if (denom != 0.0)
             {
                 intersection is;
-                is._path = vector::dot(sn, (st - ro)) / (denom);
-                is._point3 = ro + is._path * rd;
-                is._point2 = local(s, is._point3, ctx);
-                is._status = mask(is._point2.value_or(point2()));
-                is._direction = denom > 0 ? e_along : e_opposite;
+                is.path = vector::dot(sn, (st - ro)) / (denom);
+                is.point3 = ro + is.path * rd;
+                is.point2 = local(s, is.point3, ctx);
+                is.status = mask(is.point2.value_or(point2()));
+                is.direction = denom > 0 ? e_along : e_opposite;
                 return is;
             }
             return intersection{};
