+++ tests/common/include/tests/common/test_core.inl
@@ -52,14 +52,14 @@ TEST(__plugin, intersection)
     intersection i1 = {1.7, point3(0.2, 0.3, 0.), point2(0.2, 0.4), intersection_status::e_inside};
 
     intersection invalid;
-    ASSERT_TRUE(invalid._status == intersection_status::e_missed);
+    ASSERT_TRUE(invalid.status== intersection_status::e_missed);
 
     dvector<intersection> intersections = {invalid, i0, i1};
     std::sort(intersections.begin(), intersections.end());
 
-    ASSERT_NEAR(intersections[0]._path, 1.7, epsilon);
-    ASSERT_NEAR(intersections[1]._path, 2, epsilon);
-    ASSERT_TRUE(std::isinf(intersections[2]._path));
+    ASSERT_NEAR(intersections[0].path, 1.7, epsilon);
+    ASSERT_NEAR(intersections[1].path, 2, epsilon);
+    ASSERT_TRUE(std::isinf(intersections[2].path));
 }
 
 // Google Test can be run manually from the main() function
